Dev2Master_12oct22

+++ custom/addons/do_customization/models/sale.py
@@ -69,6 +69,29 @@ class SaleOrder(models.Model):
     delivering_date = fields.Datetime('Deliver Date', store=True, default="", readonly=True)
     delivered_date = fields.Datetime('Delivered Date', store=True, default="", readonly=True)
 
+    shipping_method = fields.Selection([
+        ('standard', 'Standard'),
+        ('express', 'Express')
+    ], default='standard')
+                    
+    @api.model
+    def _compute_sol_page_break(self, sol_per_page):
+        # to show limited number of sol in a page in printing invoice ( sale report template )
+        sol_total_list = []
+        sol_list = []
+        spp = sol_per_page
+        for rec in self.order_line:
+            if spp == 0:
+                sol_total_list.append(sol_list)
+                spp = sol_per_page
+                sol_list = []
+            if not rec.is_delivery:
+                sol_list.append(rec)
+                spp = spp - 1
+        if len(sol_list) > 0:
+            sol_total_list.append(sol_list)
+        return sol_total_list
+
     @api.model
     def _check_delivery_selected(self):
         delivery = 0
@@ -184,14 +207,27 @@ class SaleOrder(models.Model):
 
     def action_confirm(self):
         self.ensure_one()
-        order = self.env['sale.order.line'].search([('order_id', '=', self.id)])
-        order_copy = self.copy()
+        order_lines = self.env['sale.order.line'].search([('order_id', '=', self.id)])
+        copy_or_not = False
+        order_copy = self.env['sale.order']
+        
+        for iscopy in self.website_order_line:
+            if not iscopy.selected_checkout:
+                copy_or_not = True
+            
+        if copy_or_not:
+            order_copy = self.copy()       
+        
+        if order_copy:
+            for mma in order_copy.order_line:
+                if mma.selected_checkout or mma.is_delivery:
+                    mma.unlink()
+                    
+            for o in order_copy.website_order_line:
+                if o.selected_checkout or o.is_delivery:
+                    o.unlink()
 
-        for o in order_copy.website_order_line:
-            if o.selected_checkout or o.is_delivery:
-                o.unlink()
-
-        for o in order:
+        for o in order_lines:
             if not o.is_delivery:
                 if not o.selected_checkout:
                     o.unlink()
@@ -206,8 +242,9 @@ class SaleOrder(models.Model):
         if order_copy and self.state in ('sale','approve_by_admin'):
             self.env['website'].newlp_so_website(order_copy)
 
-        order_copy.amount_delivery = 0
-        order_copy.selected_carrier_id = ''
+        if order_copy:
+            order_copy.amount_delivery = 0
+            order_copy.selected_carrier_id = ''
 
         return res
     
@@ -317,7 +354,7 @@ class SaleOrder(models.Model):
     
         if self.write({'state': 'ready_to_pick'}):
             
-            for sol_data in self.env['sale.order.line'].search([('order_id','=',self.id)]):
+            for sol_data in self.env['sale.order.line'].search([('order_id','=',self.id),('sol_state','!=','cancel')]):
                 sol_data.write({'sol_state': 'ready_to_pick'})
                 
             picking_obj = self.env['stock.picking'].search([('origin','=',self.name)])
@@ -325,6 +362,11 @@ class SaleOrder(models.Model):
                                'ready_to_pick': True,
                                'hold_state': False })
 
+    def action_cancel(self):
+        if self.write({'state': 'cancel'}):    
+            for sol_data in self.env['sale.order.line'].search([('order_id','=',self.id)]):
+                sol_data.write({'sol_state': 'cancel'})
+                
             
 class SaleOrderLine(models.Model):
     _inherit = 'sale.order.line'
@@ -524,6 +566,7 @@ class SaleOrderLine(models.Model):
                         'amount_untaxed': sol_data.order_id.pricelist_id.currency_id.round(amount_untaxed),
                         'amount_tax': sol_data.order_id.pricelist_id.currency_id.round(amount_tax),
                         'amount_total': amount_untaxed + amount_tax,
+                        'amount_delivery': 0.0,
                     })
